requirements.txt
Django
djangorestframework
django-registration
psycopg2
django-postgres-copy
